GET /api/v1/region/room/
